Track jump level and apply jump modifier.
Every time the player jumps within `threshold` time of landing
(currently set at 1/4 of a second) the "jump level" is increased,
up until the maximum jump level.

The player's initial jumping (and falling) velocity is multiplied
by this jump modifier to provide stronger jumps.

The jump modifier is reduced for every threshold missed. If the player
jumps after 1/4 of a second has passed since landing, the jump modifier
is reduced by 1. After 1/2 of a second, reduced by 2. Etc.

The branch name is triple jump, but the jumpber of allowed jump levels
is actually 4 (0..3).